requirements.txt
asgiref==3.7.2
Django==4.2.7
factory-boy==3.3.0
Faker==20.0.3
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
